pyproject.toml
[project]
name = "mem-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "qli", email = "1787204269@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
